requirements.txt
python-dotenv
pandas
pinecone-client
langchain
openai
fastapi
uvicorn
python-multipart
starlette